Added rendering of the level boundaries
