+++ src/index.js
@@ -21,55 +21,9 @@ import {
     Level
 } from './modules/level.js'
 
-class Player {
-    constructor(spawnPosition, spawnDirection, fieldOfView, movementSpeed, rotationSpeed) {
-        this.position = spawnPosition;
-        this.direction = spawnDirection;
-        this.viewPlane = new Vector2D(0.0, Math.sin(fieldOfView * 2.0 * Math.PI / 180.0));
-        this.movementSpeed = movementSpeed;
-        this.rotationSpeed = rotationSpeed;
-    }
-
-    get position() {
-        return this._position;
-    }
-
-    set position(value) {
-        this._position = value;
-    }
-
-    get direction() {
-        return this._direction;
-    }
-
-    set direction(value) {
-        this._direction = value;
-    }
-
-    get viewPlane() {
-        return this._viewPlane;
-    }
-
-    set viewPlane(value) {
-        this._viewPlane = value;
-    }
-
-    get movementSpeed() {
-        return this._movementSpeed;
-    }
-
-    set movementSpeed(value) {
-        this._movementSpeed = value;
-    }
-
-    get rotationSpeed() {
-        return this._rotationSpeed;
-    }
-
-    set rotationSpeed(value) {
-        this._rotationSpeed = value;
-    }
-}
+import {
+    Player
+} from './modules/player.js'
 
 class PlayerController {
     constructor(player) {
@@ -126,10 +80,10 @@ class Application {
         this.renderBuffer = new RenderBuffer(document.getElementById("viewport"));
 
         //Temporary game logic variables
-        this.player = new Player(new Vector2D(100.0, 100.0), new Vector2D(-1.0, 0.0), 45, 50.0, 1.0);
+        this.player = new Player(new Vector2D(0.0, 0.0), new Vector2D(-1.0, 0.0), 45, 50.0, 1.0);
         this.playerController = new PlayerController(this.player);
 
-        this.level = new Level(new Vector2D(20, 20), this.player);
+        this.level = new Level(new Vector2D(10, 10), this.player);
     }
 
     init = (fps) => {
