+++ src/modules/level.js
@@ -10,6 +10,10 @@ import {
     Color
 } from './color.js';
 
+import {
+    Player
+} from './player.js'
+
 class Level {
     constructor(gridSize, player) {
         this.grid = new Array(gridSize.x * gridSize.y).fill(0);
@@ -36,21 +40,26 @@ class Level {
     render = (renderBuffer) => {
         const cellSize = 25;
 
-        for (let x = 0; x < this.gridSize.x; x++) {
-            const xPosition = x * cellSize;
-            if (xPosition < renderBuffer.width) {
-                drawLineDDA(new Vector2D(xPosition, 0), new Vector2D(xPosition, renderBuffer.height), new Color(0, 255, 0, 255), renderBuffer);
+        const mapScreenStartPositionX = (renderBuffer.width / 2.0) - this.player.position.x;
+        const mapScreenStartPositionY = (renderBuffer.height / 2.0) - this.player.position.y;
+        const mapScreenEndPositionX = (cellSize * this.gridSize.x) - this.player.position.x + (renderBuffer.width / 2.0);
+        const mapScreenEngPositionY = (cellSize * this.gridSize.y) - this.player.position.y + (renderBuffer.height / 2.0);
+
+        for (let x = 0; x < this.gridSize.x + 1; x++) {
+            const xPosition = mapScreenStartPositionX + (x * cellSize);
+            if (xPosition <= renderBuffer.width) {
+                drawLineDDA(new Vector2D(xPosition, mapScreenStartPositionY), new Vector2D(xPosition, mapScreenEngPositionY), new Color(0, 255, 0, 255), renderBuffer);
             }
         }
 
-        for (let y = 0; y < this.gridSize.y; y++) {
-            const yPosition = y * cellSize;
+        for (let y = 0; y < this.gridSize.y + 1; y++) {
+            const yPosition = mapScreenStartPositionY + (y * cellSize);
             if (yPosition < renderBuffer.height) {
-                drawLineDDA(new Vector2D(0, yPosition), new Vector2D(renderBuffer.width, yPosition), new Color(0, 255, 0, 255), renderBuffer);
+                drawLineDDA(new Vector2D(mapScreenStartPositionX, yPosition), new Vector2D(mapScreenEndPositionX, yPosition), new Color(0, 255, 0, 255), renderBuffer);
             }
         }
 
-        this.renderPlayer(this.player.position, renderBuffer);
+        this.renderPlayer(new Vector2D(renderBuffer.width / 2, renderBuffer.height / 2), renderBuffer);
     }
 
     renderPlayer = (screenPosition, renderTarget) => {
